requirements.txt
absl-py==0.9.0
ar-markers==0.5.0
asgiref==3.2.7
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==3.1.4
blinker==1.4
boto3==1.13.19
botocore==1.16.19
build-utils==0.3.2
cachetools==3.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
colorama==0.4.3
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.17
cytoolz==0.10.1
dask==2.17.0
declxml==0.9.1
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.6
django-rest-knox==4.1.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
entrypoints==0.3
gast==0.2.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
idna==2.9
imageio==2.8.0
imgaug==0.4.0
importlib-metadata==1.5.0
imutils==0.5.3
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
JPype1==0.7.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyter-tensorboard==0.2.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==21.2.1
kiwisolver==1.2.0
konlpy==0.5.2
lxml==4.5.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mysqlclient==1.3.14
natsort==7.0.1
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
nose==1.3.7
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.2.0.34
opt-einsum==3.1.0
pandas==1.0.4
pandocfilters==1.4.2
parso==0.7.0
pascal-voc-writer==0.1.4
pickleshare==0.7.5
Pillow==7.1.2
pkginfo==1.5.0.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyreadline==2.1
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==18.1.1
qtconsole==4.7.4
QtPy==1.9.0
readme-renderer==26.0
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.0
s3transfer==0.3.3
scikit-image==0.16.2
scipy==1.4.1
Send2Trash==1.5.0
Shapely==1.7.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
tensorboard==1.15
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toolz==0.10.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
tweepy==3.8.0
twine==3.1.1
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0
